refactor: 대시보드 - 파워유저 구조 변경
